+++ NEWS.md
@@ -7,6 +7,9 @@ New features (Analysis):
 + Don't compare parameter types against alternate method signatures which have too many required parameters.
   (e.g. warn about `max([])` but not `max([], [1])`)
 + Support `/** @unused-param $param_name */` in doc comments as an additional way to support suppressing warnings about individual parameters being unused.
++ Warn about loop conditions that potentially don't change due to the body of the loop.
+  This check uses heuristics and is prone to false positives.
+  New issue types: `PhanPossiblyInfiniteLoop`
 
 Plugins:
 + Warn about `#` comments in `PHPDocInWrongCommentPlugin` if they're not used for the expected `#[` syntax of php 8.0 attributes.